Start from a clean state
As we are reviving this repository, the old idea of having a bespoke
tool for generating Nickel contracts from openAPI specs has been
abandoned. Instead, we plan to rely on json-schema-to-nickel together
with existing external repositories that have up-to-date conversions of
the openAPI K8S specs to JSON schema.

Thus, this PR cleans many files that aren't relevant anymore, and update
the README to reflect the new status of the project.