Add contracts to flake output
This commit adds an output to generate Nickel contracts for Kubernetes
resources for the latest (known) version of Kubernetes JSON schemas. It
basically pipes together Kubernetes JSON schemas and
json-schema-to-nickel via Nix.

WARNING: this is currently not working, but not because of
nickel-kubernetes. The K8s schemas are hitting some limitations of
json-schema-to-nickel (basically unimplemented JSON schema constructs).
The proposed change is still meaningful and self-contained, so let's
push this before starting to investigate what's missing in
json-schema-to-nickel.