Add json-schema-bundler
This commit adds a small wrapper NPM package which uses
json-schema-ref-parser to take an input schema and inline all external
references (not exactly: it brings the definition in the schema and
creates internal references, which preserves sharing and cycles) so that
they can be handled by json-schema-to-nickel which currently doesn't
support external references.

The .gitignore is updated, and cleaned at this occasion.