Add contract generation
Use json-schema-bundler and a pre-processing steps to make all refs
local in order to generate the contracts for the latest version of
kubernetes-json-schema directly fro the flake.

There is still an issue that the references bundled by
json-schema-bundler have the shape "/#properties" which is currently not
handled properly by json-schema-to-nickel (it just bails out with a
warning), so the contract are expected to be highly incomplete. But it's
a first working step of end-to-end contract generation.