Do not inline js2n lib in latest contracts
The contracts from v1.31.0 to v1.34.0 have been re-generated recently.
Unfortunately, they were generated with the js2n lib inlined, which
wastes a lot of space (and isn't the process documented in the README).
This commit fixes the issue by regenerating those contracts with the
proper `--library-path` argument provided to json-schema-to-nickel, such
that the library is shared in `.js2n-lib` for each contract bundle.